feat: 히든 케이스별 소요시간 반환 로직 추가
