+++ run/src/main/java/ide/run/controller/RunRequestController.java
@@ -1,10 +1,9 @@
 package ide.run.controller;
 
-import com.amazonaws.services.lambda.runtime.events.S3ObjectLambdaEvent;
 import ide.run.domain.RequestDto;
 import ide.run.domain.ResponseDto;
-import ide.run.service.FileIOService;
-import ide.run.service.GraderService;
+import ide.run.service.fileIO.FileIOService;
+import ide.run.service.grader.GraderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -22,6 +21,8 @@ public class RunRequestController implements Function<RequestDto, ResponseDto> {
     public ResponseDto apply(RequestDto requestDto) {
         String extension = "." + requestDto.getLanguage();
         File file = fileIOService.makeFileToLambdaMemory(requestDto.getRequestCode(), requestDto.getQuestionId(), extension);
+
+
         return graderService.grader(file, requestDto);
     }
 }
