--- run/src/main/java/ide/run/exception/compile/CompileErrorException.java
@@ -1,7 +1,0 @@
-package ide.run.exception.compile;
-
-public class CompileErrorException extends RuntimeException {
-    public CompileErrorException(String message) {
-        super(message);
-    }
-}
