+++ run/src/main/java/ide/run/repository/S3Repository.java
@@ -2,15 +2,11 @@ package ide.run.repository;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.GetObjectRequest;
-import com.amazonaws.services.s3.model.S3Object;
-import com.amazonaws.services.s3.model.S3ObjectInputStream;
-import ide.run.exception.s3.S3IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import java.io.File;
-import java.io.IOException;
 
 @Component
 @RequiredArgsConstructor
@@ -30,26 +26,10 @@ public class S3Repository {
         amazonS3.putObject(bucket, PATH_USER + uuid + SLASH + file.getName(), file);
     }
 
-    public String getUserCode(String uuid, String questionId, String extension) {
-        S3Object findFile = amazonS3.getObject(bucket, PATH_USER + uuid + SLASH + questionId + extension);
-        try {
-            return ByteToString(findFile);
-        } catch (IOException e) {
-            log.info("Error by: Transfer ByteStream to String Code", e);
-            throw new S3IOException(e);
-        }
-    }
-
     public void loadAdminFile(String questionId, String fileName) {
         GetObjectRequest getObjectRequest = new GetObjectRequest(bucket, PATH_ADMIN + questionId + SLASH + fileName);
         File file = new File(PATH_LAMBDA + fileName);
         amazonS3.getObject(getObjectRequest, file);
     }
 
-    private static String ByteToString(S3Object findFile) throws IOException{
-        S3ObjectInputStream objectContent = findFile.getObjectContent();
-        byte[] bytes = objectContent.readAllBytes();
-        String StringCode = new String(bytes, "UTF-8");
-        return StringCode;
-    }
 }
